move setupHeaderFile to fix lint rants
